requirements.txt
pandas>=1.5.0
numpy>=1.21.0
streamlit>=1.25.0
dtaidistance>=2.3.0
polygon-api-client>=1.12.0
yfinance>=0.2.0
requests>=2.28.0
matplotlib>=3.5.0
mplfinance>=0.12.0
altair>=5.0.0
beautifulsoup4>=4.11.0
spacy>=3.5.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
